- Added header and version, cleanuped up output text - Incorporated better grep commands (from jumpsmm7) to handle files with comments AFTER the domain - Incorporated output to show file number being downloaded (from jumpsmm7)

+++ adblock/gen_adblock.sh
@@ -10,16 +10,23 @@
 # list format can be one of any:
 #		domains - for lists of the format of 1 domain per line
 #		hosts - for lists of the format of 1 host file entry per line with 0.0.0.0 IP (sorry, no 127.0.0.1)
-# @juched - v1.0.4
+# @juched
 # 	Special thanks to @Martineau @rgnldo @Jack_Yaz for setting up and hosting and thinking of this
 # v1.0.1 - moved config to /opt/share/unbound/configs
 #	 - save and reload unbound cache on restart
 # v1.0.2 - separated required domains from user editable domains for allow list
 # v1.0.3 - switched to use unbound_manager.sh restart to be more safe
 # v1.0.4 - cleanup and split of sites file by jumpsmm7 (Thank you!)
+# v1.0.5 - Output version header, add count of file processed, and support files with comments AFTER the domain (thanks jumpsmm7)
 
-destinationIP="0.0.0.0"
-
+echo "                             "
+echo " _____   _ _   _         _   "
+echo "|  _  |_| | |_| |___ ___| |_ "
+echo "|     | . | . | | . |  _| '_|"
+echo "|__|__|___|___|_|___|___|_,_|"
+echo " @juched - v1.0.5 - Thanks to @SomeWhereOverTheRainBow"
+echo ""
+                             
 #adblock function paths
 tempoutlist="/opt/var/lib/unbound/adblock/adlist"
 tempwhitelistoutlist="/opt/var/lib/unbound/adblock/whitelist"
@@ -52,22 +59,28 @@ fi
 download_file () {
   sites=$1
   list=$2
+  count=1
   while read -r line
   do
   set -- $line
   for url in $(echo $line); do
-  echo "Attempting to Download $url"
-  curl --progress-bar $url | grep -v "#" | grep -v "::1" | grep -v "0.0.0.0 0.0.0.0" | sed '/^$/d' | sed 's/\ /\\ /g' | awk '{print $NF}' | grep -v '^\\' | grep -v '\\$'| sort >> $list
+  echo "Attempting to Download $count of $(wc -l < $sites) from $url."
+  curl --progress-bar $url | grep -o '^[^#]*' | grep -v "::1" | grep -v "0.0.0.0 0.0.0.0" | sed '/^$/d' | sed 's/\ /\\ /g' | awk '{print $NF}' | grep -o '^[^\\]*' | grep -o '^[^\\$]*' | sort >> $list
   dos2unix $list
   done
+  count=$((count + 1))
   done < "$sites"
 }
 
 filter_file () {
   filter=$1
   original=$2
-  awk 'NR==FNR{a[$0];next} !($0 in a) {print $NF}' $filter $original | sort -u > ${original}.tmp
-  mv ${original}.tmp $original
+  if [ "$(wc -l < $filter)" -ne "0" ]; then
+    awk 'NR==FNR{a[$0];next} !($0 in a) {print $NF}' $filter $original | sort -u > ${original}.tmp
+    mv ${original}.tmp $original
+  else
+    echo "No filtering from $filter required for $original..."
+  fi
 }
 
 cleanup () { 
@@ -85,19 +98,19 @@ cleanup () {
   cat $output | sed -r -e 's/[[:space:]]+/\t/g' | sed -e 's/\t*#.*$//g' | sed -e 's/[^a-zA-Z0-9\.\_\t\-]//g' | sed -e 's/\t$//g' | sed -e '/^#/d' | sed -e 's/^[ \t]*//;s/[ \t]*$//' | sort -u | sed '/^$/d' > $3
 }
 
+echo "Downloading list(s) from block site(s) configured..."
 download_file $blocksites $tempoutlist
-[ -f $allowsites ] && echo "Downloading User Allow List..." && download_file $allowsites $tempwhitelistoutlist
-[ -f $blocklist ] && echo "Combining User Custom block host..." && cat $blocklist >> $tempoutlist
-[ -f $tempwhitelistoutlist ] && echo "Removing any downloaded whitelist items..." && filter_file $tempwhitelistoutlist $tempoutlist
-[ -f $allowlist ] && echo "Filtering user requested domains from adblock list..." && filter_file $allowlist $tempoutlist
-[ -f $permlist ] && echo "Filtering required domains from adblock list..." && filter_file $permlist $tempoutlist
+[ -f $allowsites ] && echo "Downloading list(s) from allow site(s) configured..." && download_file $allowsites $tempwhitelistoutlist
+[ -f $blocklist ] && echo "Adding user requested hosts to list..." && cat $blocklist >> $tempoutlist
+[ -f $tempwhitelistoutlist ] && echo "Removing allow list downloaded whitelist hosts from list..." && filter_file $tempwhitelistoutlist $tempoutlist
+[ -f $allowlist ] && echo "Removing user requested hosts from list..." && filter_file $allowlist $tempoutlist
+[ -f $permlist ] && echo "Removing required hosts from list..." && filter_file $permlist $tempoutlist
 [ -f $tempoutlist ] && echo "Removing unnecessary formatting from the domain list..." && cleanup $tempoutlist $outlist $finalist
-numberOfAdsBlocked=$(wc -l < $outlist)
-echo "$numberOfAdsBlocked domains compiled"
 echo "Generating Unbound adlist....."
 awk '{print "local-zone: \""$1"\" always_nxdomain"}' $finalist > $adlist
-numberOfAdsBlocked=$(wc -l < $adlist)
-echo " Number of adblocked (ads/malware/tracker) and blacklisted domains: $numberOfAdsBlocked" > $statsFile
+numberOfHostsBlocked=$(wc -l < $adlist)
+echo "Number of adblocked hosts: $numberOfHostsBlocked"
+echo " Number of adblocked (ads/malware/tracker) and blacklisted hosts: $numberOfHostsBlocked" > $statsFile
 echo " Last updated: $(date +"%c")" >> $statsFile
 
 echo "Removing temporary files..."
